Move parsing logic for call parameter sets into a separate class.
